interp/mintops.c: In function 'mono_interp_dis_mintop': interp/mintops.c:87:37: warning: format '%lld' expects argument of type 'long long int', but argument 3 has type 'long int' [-Wformat=]    g_string_append_printf (str, " %lld", (gint64)READ64 (ip + 1));
Commit migrated from https://github.com/mono/mono/commit/b504d7e6d89c94ae7b00145afd3ff0b06813553b

+++ src/mono/mono/mini/interp/mintops.c
@@ -84,7 +84,7 @@ mono_interp_dis_mintop(const guint16 *base, const guint16 *ip)
 		g_string_append_printf (str, " %d", (gint32)READ32 (ip + 1));
 		break;
 	case MintOpLongInt:
-		g_string_append_printf (str, " %lld", (gint64)READ64 (ip + 1));
+		g_string_append_printf (str, " %lld", (long long)READ64 (ip + 1));
 		break;
 	case MintOpFloat: {
 		gint32 tmp = READ32 (ip + 1);
